Adding project file to demonstrate clustering
Unsupervised machine learning Project on clustering by various methods such as KMeans and Hierarchial method. This also includes scoring the methods to determine best fit for number of clusters and the best clustering algorithm. The internal validation methods used for judging the cluster quality are: Calinski-Harbasz score and Silhouette score.